implementing comparable for connectionObject
Using remote debugging, it was clear that there was an issue with the
ChatServer.removeClientConnection() method where the main body of the
method was never executed. This was attributed to the fact that when the
connectedClients.contains(clientConnectionObject) evaluated, the
ClientConnectionObject class didn't immplement Comparable<T> and so
wasn't able to evaluate.